Remove duplicate stories
