+++ storybook/stories/Examples/AreaChart.stories.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { curveCardinal } from 'd3-shape';
 import { Args } from '@storybook/react-vite';
 import { pageData, rangeData, timeData } from '../data';
 import {
@@ -26,35 +25,6 @@ export default {
   argTypes: AreaChartArgs,
 };
 
-export const Simple = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <AreaChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Area type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
-          <Tooltip />
-          <RechartsHookInspector />
-        </AreaChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(AreaChartArgs),
-    width: 500,
-    height: 400,
-    data: pageData,
-    margin: {
-      top: 10,
-      right: 30,
-      left: 0,
-      bottom: 0,
-    },
-  },
-};
-
 export const StackedAreaChart = {
   render: (args: Args) => {
     return (
@@ -77,178 +47,6 @@ export const StackedAreaChart = {
     width: 500,
     height: 400,
     data: timeData,
-    margin: {
-      top: 10,
-      right: 30,
-      left: 0,
-      bottom: 0,
-    },
-  },
-};
-
-export const PercentAreaChart = {
-  render: (args: Args) => {
-    const toPercent = (decimal: number, fixed = 0) => `${(decimal * 100).toFixed(fixed)}%`;
-
-    const getPercent = (value: number, total: number = 0) => {
-      const ratio = total > 0 ? value / total : 0;
-
-      return toPercent(ratio, 2);
-    };
-
-    const renderTooltipContent = (o: { payload?: ReadonlyArray<{ value?: number }>; label?: string | number }) => {
-      const { payload, label } = o;
-      const total = payload?.reduce((result, entry) => result + (entry.value ?? 0), 0);
-
-      return (
-        <div className="customized-tooltip-content">
-          <p className="total">{`${label} (Total: ${total})`}</p>
-          <ul className="list">
-            {payload?.map((entry: any) => (
-              <li key={`item-${entry.name}`} style={{ color: entry.color }}>
-                {`${entry.name}: ${entry.value}(${getPercent(entry.value, total)})`}
-              </li>
-            ))}
-          </ul>
-        </div>
-      );
-    };
-
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <AreaChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis tickFormatter={toPercent} />
-          <Area type="monotone" dataKey="uv" stackId="1" stroke="#8884d8" fill="#8884d8" />
-          <Area type="monotone" dataKey="pv" stackId="1" stroke="#82ca9d" fill="#82ca9d" />
-          <Area type="monotone" dataKey="amt" stackId="1" stroke="#ffc658" fill="#ffc658" />
-          <Tooltip content={renderTooltipContent} defaultIndex={3} active />
-          <RechartsHookInspector />
-        </AreaChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(AreaChartArgs),
-    width: 500,
-    height: 400,
-    data: pageData,
-    stackOffset: 'expand',
-    margin: {
-      top: 10,
-      right: 30,
-      left: 20,
-      bottom: 20,
-    },
-  },
-};
-
-export const CardinalAreaChart = {
-  render: (args: Args) => {
-    const cardinal = curveCardinal.tension(0.2);
-
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <AreaChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Area type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" fillOpacity={0.3} />
-          <Area type={cardinal} dataKey="uv" stroke="#82ca9d" fill="#82ca9d" fillOpacity={0.3} />
-          <Tooltip />
-          <RechartsHookInspector />
-        </AreaChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(AreaChartArgs),
-    width: 500,
-    height: 400,
-    data: pageData,
-    margin: {
-      top: 10,
-      right: 30,
-      left: 0,
-      bottom: 0,
-    },
-  },
-};
-
-export const AreaChartConnectNulls = {
-  render: (args: Args) => {
-    return (
-      <div style={{ width: '100%' }}>
-        <ResponsiveContainer width="100%" height={200}>
-          <AreaChart {...args}>
-            <CartesianGrid strokeDasharray="3 3" />
-            <XAxis dataKey="name" />
-            <YAxis />
-            <Area type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
-            <Tooltip />
-          </AreaChart>
-        </ResponsiveContainer>
-        <ResponsiveContainer width="100%" height={200}>
-          <AreaChart {...args}>
-            <CartesianGrid strokeDasharray="3 3" />
-            <XAxis dataKey="name" />
-            <YAxis />
-            <Area connectNulls type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
-            <Tooltip />
-            <RechartsHookInspector />
-          </AreaChart>
-        </ResponsiveContainer>
-      </div>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(AreaChartArgs),
-    width: 500,
-    height: 200,
-    data: [
-      {
-        name: 'Page A',
-        uv: 4000,
-        pv: 2400,
-        amt: 2400,
-      },
-      {
-        name: 'Page B',
-        uv: 3000,
-        pv: 1398,
-        amt: 2210,
-      },
-      {
-        name: 'Page C',
-        uv: 2000,
-        pv: 9800,
-        amt: 2290,
-      },
-      {
-        name: 'Page D',
-        pv: 3908,
-        amt: 2000,
-      },
-      {
-        name: 'Page E',
-        uv: 1890,
-        pv: 4800,
-        amt: 2181,
-      },
-      {
-        name: 'Page F',
-        uv: 2390,
-        pv: 3800,
-        amt: 2500,
-      },
-      {
-        name: 'Page G',
-        uv: 3490,
-        pv: 4300,
-        amt: 2100,
-      },
-    ],
     margin: {
       top: 10,
       right: 30,
@@ -325,153 +123,6 @@ export const StackedAreaChartConnectNulls = {
       {
         name: 'Page F',
         uv: 2390,
-        pv: 3800,
-        amt: 2500,
-      },
-      {
-        name: 'Page G',
-        uv: 3490,
-        pv: 4300,
-        amt: 2100,
-      },
-    ],
-    margin: {
-      top: 10,
-      right: 30,
-      left: 0,
-      bottom: 0,
-    },
-  },
-};
-
-export const SynchronisedAreaChart = {
-  render: (args: Args) => {
-    return (
-      <div style={{ width: '100%' }}>
-        <div style={{ width: '100%', display: 'flex', flexDirection: 'row' }}>
-          <ResponsiveContainer width="30%" height={200}>
-            <AreaChart {...args}>
-              <CartesianGrid strokeDasharray="3 3" />
-              <XAxis dataKey="name" />
-              <YAxis />
-              <Area type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
-              <Tooltip />
-              <RechartsHookInspector />
-            </AreaChart>
-          </ResponsiveContainer>
-          <ResponsiveContainer width="70%" height={200}>
-            <AreaChart {...args}>
-              <CartesianGrid strokeDasharray="3 3" />
-              <XAxis dataKey="name" />
-              <YAxis />
-              <Area type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
-              <Tooltip />
-              <RechartsHookInspector />
-            </AreaChart>
-          </ResponsiveContainer>
-        </div>
-        <ResponsiveContainer width="100%" height={200}>
-          <AreaChart {...args}>
-            <CartesianGrid strokeDasharray="3 3" />
-            <XAxis dataKey="name" />
-            <YAxis />
-            <Area type="monotone" dataKey="pv" stroke="#82ca9d" fill="#82ca9d" />
-            <Tooltip />
-          </AreaChart>
-        </ResponsiveContainer>
-      </div>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(AreaChartArgs),
-    width: 500,
-    height: 200,
-    data: pageData,
-    margin: {
-      top: 10,
-      right: 30,
-      left: 0,
-      bottom: 0,
-    },
-    syncId: 'anyId',
-  },
-};
-
-export const AreaChartFillByValue = {
-  render: (args: Args) => {
-    const gradientOffset = () => {
-      const dataMax = Math.max(...args.data.map((i: any) => i.uv));
-      const dataMin = Math.min(...args.data.map((i: any) => i.uv));
-
-      if (dataMax <= 0) {
-        return 0;
-      }
-      if (dataMin >= 0) {
-        return 1;
-      }
-
-      return dataMax / (dataMax - dataMin);
-    };
-
-    const off = gradientOffset();
-
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <AreaChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <defs>
-            <linearGradient id="splitColor" x1="0" y1="0" x2="0" y2="1">
-              <stop offset={off} stopColor="green" stopOpacity={1} />
-              <stop offset={off} stopColor="red" stopOpacity={1} />
-            </linearGradient>
-          </defs>
-          <Area type="monotone" dataKey="uv" stroke="#000" fill="url(#splitColor)" />
-          <Tooltip />
-          <RechartsHookInspector />
-        </AreaChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(AreaChartArgs),
-    width: 500,
-    height: 400,
-    data: [
-      {
-        name: 'Page A',
-        uv: 4000,
-        pv: 2400,
-        amt: 2400,
-      },
-      {
-        name: 'Page B',
-        uv: 3000,
-        pv: 1398,
-        amt: 2210,
-      },
-      {
-        name: 'Page C',
-        uv: -1000,
-        pv: 9800,
-        amt: 2290,
-      },
-      {
-        name: 'Page D',
-        uv: 500,
-        pv: 3908,
-        amt: 2000,
-      },
-      {
-        name: 'Page E',
-        uv: -2000,
-        pv: 4800,
-        amt: 2181,
-      },
-      {
-        name: 'Page F',
-        uv: -250,
         pv: 3800,
         amt: 2500,
       },
