+++ storybook/stories/Examples/BarChart/BarChart.stories.tsx
@@ -6,7 +6,6 @@ import {
   BarChart,
   Brush,
   CartesianGrid,
-  Cell,
   ErrorBar,
   LabelList,
   Legend,
@@ -27,37 +26,6 @@ export default {
   component: BarChart,
   docs: {
     autodocs: false,
-  },
-};
-
-export const Simple = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Legend />
-          <Tooltip defaultIndex={3} />
-          <Bar dataKey="pv" fill="#8884d8" activeBar={<Rectangle fill="pink" stroke="blue" />} />
-          <Bar dataKey="uv" fill="#82ca9d" activeBar={<Rectangle fill="gold" stroke="purple" />} />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    margin: {
-      top: 5,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-    data: pageData,
   },
 };
 
@@ -169,193 +137,6 @@ export const StackedWithErrorBar = {
   },
 };
 
-export const Mix = {
-  render: (args: Args) => {
-    const [focusedDataKey, setFocusedDataKey] = useState<string | null>(null);
-    const [locked, setLocked] = useState<boolean>(false);
-
-    const onLegendMouseEnter = (payload: LegendPayload) => {
-      if (!locked) {
-        setFocusedDataKey(String(payload.dataKey));
-      }
-    };
-
-    const onLegendMouseOut = () => {
-      if (!locked) {
-        setFocusedDataKey(null);
-      }
-    };
-
-    const onLegendClick = (payload: LegendPayload) => {
-      if (focusedDataKey === String(payload.dataKey)) {
-        if (locked) {
-          setFocusedDataKey(null);
-          setLocked(false);
-        } else {
-          setLocked(true);
-        }
-      } else {
-        setFocusedDataKey(String(payload.dataKey));
-        setLocked(true);
-      }
-    };
-
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Legend onMouseEnter={onLegendMouseEnter} onMouseOut={onLegendMouseOut} onClick={onLegendClick} />
-          <Tooltip />
-          <Bar dataKey="pv" stackId="a" fill={focusedDataKey == null || focusedDataKey === 'pv' ? '#8884d8' : '#eee'} />
-          <Bar
-            dataKey="amt"
-            stackId="a"
-            fill={focusedDataKey == null || focusedDataKey === 'amt' ? '#82ca9d' : '#eee'}
-          />
-          <Bar dataKey="uv" fill={focusedDataKey == null || focusedDataKey === 'uv' ? '#ffc658' : '#eee'} />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: pageData,
-    margin: {
-      top: 20,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
-
-export const CustomShape = {
-  render: (args: Args) => {
-    const getPath = (x: number, y: number, width: number, height: number) => {
-      return `M${x},${y + height}C${x + width / 3},${y + height} ${x + width / 2},${y + height / 3}
-  ${x + width / 2}, ${y}
-  C${x + width / 2},${y + height / 3} ${x + (2 * width) / 3},${y + height} ${x + width}, ${y + height}
-  Z`;
-    };
-
-    const TriangleBar = (props: any) => {
-      const { fill, x, y, width, height } = props;
-
-      return <path d={getPath(x, y, width, height)} stroke="none" fill={fill} />;
-    };
-
-    const colors = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042', 'red', 'pink'];
-
-    return (
-      <BarChart {...args}>
-        <CartesianGrid strokeDasharray="3 3" />
-        <XAxis dataKey="name" />
-        <YAxis />
-        <Tooltip />
-        <Bar dataKey="uv" fill="#8884d8" shape={<TriangleBar />} label={{ position: 'top' }}>
-          {pageData.map(({ name }, index) => (
-            <Cell key={`cell-${name}`} fill={colors[index % 20]} />
-          ))}
-        </Bar>
-        <RechartsHookInspector />
-      </BarChart>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: pageData,
-    margin: {
-      top: 20,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
-
-const positiveAndNegativeData = [
-  {
-    name: 'Page A',
-    uv: 4000,
-    pv: 2400,
-    amt: 2400,
-  },
-  {
-    name: 'Page B',
-    uv: -3000,
-    pv: 1398,
-    amt: 2210,
-  },
-  {
-    name: 'Page C',
-    uv: -2000,
-    pv: -9800,
-    amt: 2290,
-  },
-  {
-    name: 'Page D',
-    uv: 2780,
-    pv: 3908,
-    amt: 2000,
-  },
-  {
-    name: 'Page E',
-    uv: -1890,
-    pv: 4800,
-    amt: 2181,
-  },
-  {
-    name: 'Page F',
-    uv: 2390,
-    pv: -3800,
-    amt: 2500,
-  },
-  {
-    name: 'Page G',
-    uv: 3490,
-    pv: 4300,
-    amt: 2100,
-  },
-];
-export const PositiveAndNegative = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Legend />
-          <ReferenceLine y={0} stroke="#000" />
-          <Tooltip />
-          <Bar dataKey="pv" fill="#8884d8" />
-          <Bar dataKey="uv" fill="#82ca9d" />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: positiveAndNegativeData,
-    margin: {
-      top: 5,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
-
 const dataForBrush = [
   { name: '1', uv: 300, pv: 456 },
   { name: '2', uv: -145, pv: 230 },
@@ -397,38 +178,6 @@ const dataForBrush = [
   { name: '39', uv: -66, pv: 154 },
   { name: '40', uv: -50, pv: 186 },
 ];
-export const WithBrush = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Legend verticalAlign="top" wrapperStyle={{ lineHeight: '40px' }} />
-          <Tooltip />
-          <ReferenceLine y={0} stroke="#000" />
-          <Brush dataKey="name" height={30} stroke="#8884d8" />
-          <Bar dataKey="pv" fill="#8884d8" />
-          <Bar dataKey="uv" fill="#82ca9d" />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: dataForBrush,
-    margin: {
-      top: 5,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
 
 export const XAxisTickMarginWithBrushDy = {
   render: (args: Args) => {
@@ -463,37 +212,6 @@ export const XAxisTickMarginWithBrushDy = {
   },
 };
 
-export const WithCustomizedEvent = {
-  render: (args: Args) => {
-    const [activeIndex, setActiveIndex] = useState(1);
-
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <Bar onClick={(_data, index) => setActiveIndex(index)} dataKey="uv" isAnimationActive={false}>
-            {pageData.map(({ name }, index: number) => (
-              <Cell cursor="pointer" fill={index === activeIndex ? '#82ca9d' : '#8884d8'} key={`cell-${name}`} />
-            ))}
-          </Bar>
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: pageData,
-    margin: {
-      top: 20,
-      right: 20,
-      bottom: 20,
-      left: 20,
-    },
-  },
-};
-
 export const StackedWithBrush = {
   render: (args: Args) => {
     return (
@@ -507,102 +225,6 @@ export const StackedWithBrush = {
           <Bar dataKey="pv" stackId="a" fill="#8884d8" activeBar={{ fill: 'gold' }} />
           <Bar dataKey="uv" stackId="a" fill="#82ca9d" activeBar={{ fill: 'silver' }} />
           <Brush dataKey="name" height={30} stroke="#8884d8" />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: pageData,
-    margin: {
-      top: 5,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
-
-export const StackedBySign = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Legend />
-          <Tooltip />
-          <ReferenceLine y={0} stroke="#000" />
-          <Bar dataKey="pv" fill="#8884d8" stackId="stack" />
-          <Bar dataKey="uv" fill="#82ca9d" stackId="stack" />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: positiveAndNegativeData,
-    margin: {
-      top: 5,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-    stackOffset: 'sign',
-  },
-};
-
-export const Biaxial = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid yAxisId="left" strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis yAxisId="left" orientation="left" stroke="#8884d8" />
-          <YAxis yAxisId="right" orientation="right" stroke="#82ca9d" />
-          <Legend />
-          <Tooltip />
-          <Bar yAxisId="left" dataKey="pv" fill="#8884d8" />
-          <Bar yAxisId="right" dataKey="uv" fill="#82ca9d" />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: pageData,
-    margin: {
-      top: 20,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
-
-export const HasBackground = {
-  render: (args: Args) => {
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="name" />
-          <YAxis />
-          <Legend />
-          <Tooltip />
-          <Bar dataKey="pv" fill="#8884d8" background={{ fill: '#eee' }} />
-          <Bar dataKey="uv" fill="#82ca9d" />
           <RechartsHookInspector />
         </BarChart>
       </ResponsiveContainer>
@@ -650,149 +272,6 @@ export const HasLabelBasedOnSeparateDataKey = {
     width: 500,
     height: 300,
     data: pageData,
-    margin: {
-      top: 5,
-      right: 30,
-      left: 20,
-      bottom: 5,
-    },
-  },
-};
-
-const dataWithMultiXAxis = [
-  {
-    date: '2000-01',
-    uv: 4000,
-    pv: 2400,
-    amt: 2400,
-  },
-  {
-    date: '2000-02',
-    uv: 3000,
-    pv: 1398,
-    amt: 2210,
-  },
-  {
-    date: '2000-03',
-    uv: 2000,
-    pv: 9800,
-    amt: 2290,
-  },
-  {
-    date: '2000-04',
-    uv: 2780,
-    pv: 3908,
-    amt: 2000,
-  },
-  {
-    date: '2000-05',
-    uv: 1890,
-    pv: 4800,
-    amt: 2181,
-  },
-  {
-    date: '2000-06',
-    uv: 2390,
-    pv: 3800,
-    amt: 2500,
-  },
-  {
-    date: '2000-07',
-    uv: 3490,
-    pv: 4300,
-    amt: 2100,
-  },
-  {
-    date: '2000-08',
-    uv: 4000,
-    pv: 2400,
-    amt: 2400,
-  },
-  {
-    date: '2000-09',
-    uv: 3000,
-    pv: 1398,
-    amt: 2210,
-  },
-  {
-    date: '2000-10',
-    uv: 2000,
-    pv: 9800,
-    amt: 2290,
-  },
-  {
-    date: '2000-11',
-    uv: 2780,
-    pv: 3908,
-    amt: 2000,
-  },
-  {
-    date: '2000-12',
-    uv: 1890,
-    pv: 4800,
-    amt: 2181,
-  },
-];
-export const WithMultiXAxis = {
-  render: (args: Args) => {
-    const monthTickFormatter = (tick: any) => {
-      const date = new Date(tick);
-
-      return date.getMonth() + 1;
-    };
-
-    const renderQuarterTick = (tickProps: any) => {
-      const { x, y, payload, width, visibleTicksCount } = tickProps;
-      const { value, offset } = payload;
-      const date = new Date(value);
-      const month = date.getMonth();
-      const quarterNo = Math.floor(month / 3) + 1;
-      if (month % 3 === 1) {
-        return (
-          <text x={x + width / visibleTicksCount / 2 - offset} y={y - 4} textAnchor="middle">{`Q${quarterNo}`}</text>
-        );
-      }
-
-      const isLast = month === 11;
-
-      if (month % 3 === 0 || isLast) {
-        const pathX = Math.floor(isLast ? x - offset + width / visibleTicksCount : x - offset) + 0.5;
-
-        return <path d={`M${pathX},${y - 4}v${-35}`} stroke="red" />;
-      }
-      return null;
-    };
-
-    return (
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart {...args}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis dataKey="date" tickFormatter={monthTickFormatter as any} />
-          <XAxis
-            dataKey="date"
-            axisLine={false}
-            tickLine={false}
-            interval={0}
-            tick={renderQuarterTick}
-            height={1}
-            scale="band"
-            xAxisId="quarter"
-          />
-          <YAxis />
-          <Legend />
-          <Tooltip />
-          <Bar dataKey="pv" fill="#8884d8" />
-          <Bar dataKey="uv" fill="#82ca9d" />
-          <RechartsHookInspector />
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(BarChartArgs),
-    width: 500,
-    height: 300,
-    data: dataWithMultiXAxis,
     margin: {
       top: 5,
       right: 30,
